separated NoteImage component
